+++ docs/i18n.md
@@ -1,0 +1,11 @@
+# Internalization
+
+## Table of Contents
+
+- [Usage](#usage)
+
+- [Supported Languages](#supported-languages)
+
+- [Adding New Language](#adding-new-language)
+
+## Supported Languages
